+++ NEWS.md
@@ -1,4 +1,4 @@
-# dbx 0.2.6 [unreleased]
+# dbx 0.2.6 (2020-06-14)
 
 - Added `transaction` option to `dbxUpdate`
 - Added unsafe version check for RMySQL